requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==2.3.7
WTForms==3.1.1
python-socketio==5.10.0
python-engineio==4.8.0

# Database
SQLAlchemy==2.0.23
alembic==1.12.1

# Data processing
pandas==2.1.1
numpy==1.26.0

# API clients
anthropic==0.7.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
uuid==1.30
pytz==2023.3.post1

# Testing
pytest==7.4.2
pytest-flask==1.3.0
pytest-cov==4.1.0
coverage==7.3.2
faker==19.10.0